+++ src/lib/components/table/module-verification-details/index.tsx
@@ -4,6 +4,7 @@ import { ModulePathLink } from "../modules/ModulePathLink";
 import { TableHeader, TableRow } from "../tableComponents";
 import { useInternalNavigate } from "lib/app-provider";
 import type { MoveVerificationModuleIdentifier } from "lib/services/types";
+import { MoveVerifyStatus } from "lib/types";
 
 interface ModuleVerificationDetailsTableProps {
   moduleIdentifiers: MoveVerificationModuleIdentifier[];
@@ -39,7 +40,12 @@ export const ModuleVerificationDetailsTable = ({
             }
           >
             <TableRow>
-              <ModulePathLink hexAddr={address} moduleName={name} />
+              {/* TODO? */}
+              <ModulePathLink
+                hexAddr={address}
+                moduleName={name}
+                moveVerifyStatus={MoveVerifyStatus.Verified}
+              />
             </TableRow>
           </Grid>
         </Box>