+++ src/lib/components/table/modules/ModulesTableMobileCard.tsx
@@ -5,21 +5,29 @@ import { MobileLabel } from "../MobileLabel";
 import { useInternalNavigate } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
-import type { ModuleInfo } from "lib/types";
-import { dateFromNow, formatUTC } from "lib/utils";
+import type { MoveVerifyInfoResponse } from "lib/services/types";
+import type { ModuleInfo, Option } from "lib/types";
+import { dateFromNow, formatUTC, resolveMoveVerifyStatus } from "lib/utils";
 
 import { ModulePathLink } from "./ModulePathLink";
 
 interface ModulesTableMobileCardProps {
   moduleInfo: ModuleInfo;
+  moveVerifyInfo: Option<MoveVerifyInfoResponse>;
 }
 
 export const ModulesTableMobileCard = ({
   moduleInfo,
+  moveVerifyInfo,
 }: ModulesTableMobileCardProps) => {
   const navigate = useInternalNavigate();
   const formatAddresses = useFormatAddresses();
   const { address: creator } = formatAddresses(moduleInfo.address);
+
+  const moveVerifyStatus = resolveMoveVerifyStatus(
+    moduleInfo.digest,
+    moveVerifyInfo?.digest
+  );
 
   return (
     <MobileCardTemplate
@@ -38,6 +46,7 @@ export const ModulesTableMobileCard = ({
           <ModulePathLink
             hexAddr={moduleInfo.address}
             moduleName={moduleInfo.moduleName}
+            moveVerifyStatus={moveVerifyStatus}
           />
         </Flex>
       }